Support input (change) symbol text before search (#18)
This change is to allow user to change/input text for search. If the search is
initiated via command (F1), then the input box will be shown before search starts
to run.

This change also fix the issue that on the latest vscode, if .vscode does not exist
then this extension will help to create one.