Leon fix input symbol text (#19)
* Support input (change) symbol text before search

This change is to allow user to change/input text for search. If the search is
initiated via command (F1), then the input box will be shown before search starts
to run.

This change also fix the issue that on the latest vscode, if .vscode does not exist
then this extension will help to create one.

* fix symbol input not working

Last version enabled symbol input before search. However it's not really working as expected.
This is the fix for it.