Leon compatible with old settings (#33)
* Disable definition provider

The definition provide will run in background and slow
down the engine in many cases as spawnSync now is used
for the system. May enable it back once it change to
async.

Also udpated package to remove security vulnerabilities.

* Support to place database anywhere.

* Compatible with old version

The change for 0.0.12 somehow cause extension not working
if there is old config json file.
Adding change to set the two new fields with default value
if the current configure doesn't have them.